pyproject.toml
[project]
name = "mactag2dropbox"
version = "0.1.0"
description = "MacTag2Dropbox is a Python script that reads Finder tags from photos on macOS and uploads the photos to Dropbox, preserving their tags in Dropbox's format"
authors = [
    {name = "sago2693",email = "56797497+sago2693@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.8"
[tool.poetry.dependencies]
python = "^3.8"  # or whatever version you're using
dropbox = "^12.0.2"
macos-tags = "^1.5.1"
pandas = "^2.0.0"
pillow = "10.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry] 
package-mode = false 

